requirements/api.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
sqlalchemy[asyncio]>=2.0,<3.0
asyncpg>=0.30,<1.0
pydantic[email]>=2.10,<3.0
pydantic-settings>=2.6,<3.0
pyjwt>=2.9,<3.0
cryptography>=44.0,<45.0
httpx>=0.28,<1.0
sentry-sdk[fastapi]>=2.18,<3.0
twilio>=9.3,<10.0
openai>=1.55,<2.0
python-multipart>=0.0.17,<1.0
